+++ makefile
@@ -33,19 +33,13 @@ define cmake_build_target_perf
  	cd build-release-debug && cmake -GNinja -Wno-dev -DNO_WALL=ON -DCMAKE_BUILD_TYPE=RelWithDebInfo $(CMAKE_FLAGS) --target -DBUILD_TESTS=ON .. && ninja $(NINJA_FLAGS) $1 install
 endef
 
-define communication_generation
-	cd soccer/src/soccer/strategy/agent/communication && python3 communication_gen.py
-endef
-
 all-perf:
-	$(call communication_generation, all)
 	$(call cmake_build_target_perf, all)
 # perf (or "RelWithDebInfo"): almost as fast as release, some debug symbols
 perf: all-perf
 
 # used in GH Actions build-and-test
 all:
-	$(call communication_generation, all)
 	$(call cmake_build_target, all)
 # debug: slow executable, but many debug symbols (e.g. for GDB)
 debug: all
@@ -63,7 +57,6 @@ release: all-release
 # run if build-release-debug/ exists from a previous build
 # and no CMake files or launch.py files have been changed
 again:
-	$(call communication_generation, all)
 	(cd build-release-debug/ && ninja install)
 
 # run soccer with default flags
